feat: add key extractor to select component
