+++ components/UI/Select/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useSelect } from './useSelect';
+export { useSelectOptions } from './useSelectOptions';
