+++ components/UI/Select/hooks/useSelectOptions.tsx
@@ -3,12 +3,12 @@ import { UseSelectOptionsParams } from '../type';
 export const useSelectOptions = <Option,>({
   options,
   getLabel,
+  keyExtractor,
 }: UseSelectOptionsParams<Option>) => {
   return (
     <>
       {options.map((item: Option, index) => (
-        // eslint-disable-next-line react/no-array-index-key
-        <option key={index} value={index}>
+        <option key={keyExtractor(item)} value={index}>
           {getLabel(item)}
         </option>
       ))}
