+++ components/UI/Select/type.ts
@@ -1,6 +1,7 @@
 export type UseSelectOptionsParams<Option> = {
   options: readonly Option[];
   getLabel: (option: Option) => string;
+  keyExtractor: (option: Option) => string;
 };
 export type UseSelectParams<Option> = {
   selectedOption?: Option;
