--- components/UI/SelectItem.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SelectItem = () => {
-  return <div />;
-};
-
-export default SelectItem;
