+++ components/UI/index.ts
@@ -4,7 +4,6 @@ export { default as Input } from './Input';
 export { default as Column } from './layouts/Column';
 export { default as Flex } from './layouts/Flex';
 export { default as Grid } from './layouts/Grid';
-export { default as List } from './List';
-export { default as ListItem } from './ListItem';
+export { default as List } from './List/List';
+export { default as ListItem } from './List/ListItem';
 export { default as Select } from './Select/Select';
-export { default as SelectItem } from './SelectItem';
